Refactor EventBuffer to a top level class
